requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.6
django-crispy-forms==2.1
djangorestframework==3.15.1
kombu==5.3.7
prompt_toolkit==3.0.45
psycopg==3.1.19
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.4
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
